create SelectPenType
